Library: restructure componentes
