+++ src/faebryk/library/library/components.py
@@ -3,9 +3,9 @@
 
 import logging
 from typing import List
+from faebryk.library.library.electronic.base.nand import NAND
 
 from faebryk.library.trait_impl.component import (
-    can_bridge_defined,
     has_defined_footprint,
     has_defined_type_description,
     has_symmetric_footprint_pinmap,
@@ -13,273 +13,16 @@ from faebryk.library.trait_impl.component import (
 from faebryk.library.traits.component import (
     contructable_from_component,
     has_footprint_pinmap,
-    has_type_description,
 )
 from faebryk.library.traits.interface import contructable_from_interface_list
 from faebryk.libs.util import consume_iterator
 
-logger = logging.getLogger("library")
 
-from faebryk.library.core import Component, ComponentTrait, Parameter
+from faebryk.library.core import Component, ComponentTrait
 from faebryk.library.library.interfaces import Electrical, Power
-from faebryk.library.library.parameters import Constant
-from faebryk.library.util import times, unit_map
+from faebryk.library.util import times
 
-
-class Resistor(Component):
-    def _setup_traits(self):
-        # class _contructable_from_component(contructable_from_component.impl()):
-        #    @staticmethod
-        #    def from_component(comp: Component, resistance: Parameter) -> Resistor:
-        #        interfaces = comp.IFs.get_all()
-        #        assert len(interfaces) == 2
-        #        assert len([i for i in interfaces if type(i) is not Electrical]) == 0
-
-        #        r = Resistor.__new__(Resistor)
-        #        r.set_resistance(resistance)
-        #        class _IFs(Component.InterfacesCls()):
-        #            unnamed = interfaces
-
-        #        r.IFs = _IFs(r)
-
-        #        return r
-
-        # self.add_trait(_contructable_from_component())
-        pass
-
-    def _setup_interfaces(self):
-        class _IFs(Component.InterfacesCls()):
-            unnamed = times(2, Electrical)
-
-        self.IFs = _IFs(self)
-
-        self.add_trait(can_bridge_defined(*self.IFs.unnamed))
-
-    def __new__(cls, *args, **kwargs):
-        self = super().__new__(cls)
-        self._setup_traits()
-        return self
-
-    def __init__(self, resistance: Parameter):
-        super().__init__()
-
-        self._setup_interfaces()
-        self.set_resistance(resistance)
-
-    def set_resistance(self, resistance: Parameter):
-        self.resistance = resistance
-
-        if type(resistance) is not Constant:
-            # TODO this is a bit ugly
-            # it might be that there was another more abstract valid trait
-            # but this challenges the whole trait overriding mechanism
-            # might have to make a trait stack thats popped or so
-            self.del_trait(has_type_description)
-            return
-
-        class _has_type_description(has_type_description.impl()):
-            @staticmethod
-            def get_type_description():
-                assert isinstance(self.resistance, Constant)
-                resistance: Constant = self.resistance
-                return unit_map(
-                    resistance.value, ["µΩ", "mΩ", "Ω", "KΩ", "MΩ", "GΩ"], start="Ω"
-                )
-
-        self.add_trait(_has_type_description())
-
-
-class Capacitor(Component):
-    def __new__(cls, *args, **kwargs):
-        self = super().__new__(cls)
-        self._setup_traits()
-        return self
-
-    def __init__(self, capacitance: Parameter):
-        super().__init__()
-
-        self._setup_interfaces()
-        self.set_capacitance(capacitance)
-
-    def _setup_traits(self):
-        pass
-
-    def _setup_interfaces(self):
-        class _IFs(Component.InterfacesCls()):
-            unnamed = times(2, Electrical)
-
-        self.IFs = _IFs(self)
-        self.add_trait(can_bridge_defined(*self.IFs.unnamed))
-
-    def set_capacitance(self, capacitance: Parameter):
-        self.capacitance = capacitance
-
-        if type(capacitance) is not Constant:
-            return
-        _capacitance: Constant = capacitance
-
-        class _has_type_description(has_type_description.impl()):
-            @staticmethod
-            def get_type_description():
-                capacitance = self.capacitance
-                return unit_map(
-                    _capacitance.value, ["µF", "mF", "F", "KF", "MF", "GF"], start="F"
-                )
-
-        self.add_trait(_has_type_description())
-
-
-class BJT(Component):
-    def __new__(cls):
-        self = super().__new__(cls)
-        self._setup_traits()
-        return self
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._setup_interfaces()
-
-    def _setup_traits(self):
-        self.add_trait(has_defined_type_description("BJT"))
-
-    def _setup_interfaces(self):
-        class _IFs(Component.InterfacesCls()):
-            emitter = Electrical()
-            base = Electrical()
-            collector = Electrical()
-
-        self.IFs = _IFs(self)
-
-
-class MOSFET(Component):
-    def __new__(cls):
-        self = super().__new__(cls)
-        self._setup_traits()
-        return self
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._setup_interfaces()
-
-    def _setup_traits(self):
-        self.add_trait(has_defined_type_description("MOSFET"))
-
-    def _setup_interfaces(self):
-        class _IFs(Component.InterfacesCls()):
-            source = Electrical()
-            gate = Electrical()
-            drain = Electrical()
-
-        self.IFs = _IFs(self)
-
-
-class LED(Component):
-    class has_calculatable_needed_series_resistance(ComponentTrait):
-        @staticmethod
-        def get_needed_series_resistance_ohm(input_voltage_V: float) -> Constant:
-            raise NotImplemented
-
-    def _setup_traits(self):
-        self.add_trait(has_defined_type_description("LED"))
-
-    def _setup_interfaces(self):
-        class _IFs(Component.InterfacesCls()):
-            anode = Electrical()
-            cathode = Electrical()
-
-        self.IFs = _IFs(self)
-
-    def __new__(cls):
-        self = super().__new__(cls)
-        self._setup_traits()
-        return self
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._setup_interfaces()
-
-    def set_forward_parameters(self, voltage_V: Parameter, current_A: Parameter):
-        if type(voltage_V) is Constant and type(current_A) is Constant:
-            _voltage_V: Constant = voltage_V
-            _current_A: Constant = current_A
-
-            class _(self.has_calculatable_needed_series_resistance.impl()):
-                @staticmethod
-                def get_needed_series_resistance_ohm(
-                    input_voltage_V: float,
-                ) -> Constant:
-                    return LED.needed_series_resistance_ohm(
-                        input_voltage_V, _voltage_V.value, _current_A.value
-                    )
-
-            self.add_trait(_())
-
-    @staticmethod
-    def needed_series_resistance_ohm(
-        input_voltage_V: float, forward_voltage_V: float, forward_current_A: float
-    ) -> Constant:
-        return Constant(int((input_voltage_V - forward_voltage_V) / forward_current_A))
-
-
-class Potentiometer(Component):
-    def __new__(cls, *args, **kwargs):
-        self = super().__new__(cls, *args, **kwargs)
-        self._setup_traits()
-        return self
-
-    def __init__(self, resistance: Parameter) -> None:
-        super().__init__()
-        self._setup_interfaces(resistance)
-
-    def _setup_traits(self):
-        pass
-
-    def _setup_interfaces(self, resistance):
-        class _IFs(Component.InterfacesCls()):
-            resistors = times(2, Electrical)
-            wiper = Electrical()
-
-        class _CMPs(Component.ComponentsCls()):
-            resistors = [Resistor(resistance) for _ in range(2)]
-
-        self.IFs = _IFs(self)
-        self.CMPs = _CMPs(self)
-
-        self.IFs.wiper.connect_all(
-            [
-                self.CMPs.resistors[0].IFs.unnamed[1],
-                self.CMPs.resistors[1].IFs.unnamed[1],
-            ]
-        )
-
-        for i, resistor in enumerate(self.CMPs.resistors):
-            self.IFs.resistors[i].connect(resistor.IFs.unnamed[0])
-
-    def connect_as_voltage_divider(self, high, low, out):
-        self.IFs.resistors[0].connect(high)
-        self.IFs.resistors[1].connect(low)
-        self.IFs.wiper.connect(out)
-
-
-class Switch(Component):
-    def _setup_traits(self):
-        self.add_trait(has_defined_type_description("SW"))
-
-    def _setup_interfaces(self):
-        class _IFs(Component.InterfacesCls()):
-            unnamed = times(2, Electrical)
-
-        self.IFs = _IFs(self)
-        self.add_trait(can_bridge_defined(*self.IFs.unnamed))
-
-    def __new__(cls):
-        self = super().__new__(cls)
-        self._setup_traits()
-        return self
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._setup_interfaces()
+logger = logging.getLogger("library")
 
 
 class PJ398SM(Component):
@@ -302,59 +45,6 @@ class PJ398SM(Component):
             switch = Electrical()
 
         self.IFs = _IFs(self)
-
-
-class NAND(Component):
-    def _setup_traits(self):
-        class _constructable_from_component(contructable_from_component.impl()):
-            @staticmethod
-            def from_comp(comp: Component) -> NAND:
-                n = NAND.__new__(NAND)
-                n.__init_from_comp(comp)
-                return n
-
-        self.add_trait(_constructable_from_component())
-
-    def _setup_interfaces(self, input_cnt):
-        class _IFs(Component.InterfacesCls()):
-            power = Power()
-            output = Electrical()
-            inputs = times(input_cnt, Electrical)
-
-        self.IFs = _IFs(self)
-
-    def __new__(cls, *args, **kwargs):
-        self = super().__new__(cls)
-
-        self._setup_traits()
-
-        return self
-
-    def __init__(self, input_cnt: int):
-        super().__init__()
-
-        self._setup_interfaces(input_cnt)
-
-    def __init_from_comp(self, comp: Component):
-        interfaces = comp.IFs.get_all()
-        assert all(map(lambda i: type(i) is Electrical, interfaces))
-
-        it = iter(interfaces)
-
-        self.IFs.power = (
-            Power().get_trait(contructable_from_interface_list).from_interfaces(it)
-        )
-        self.IFs.output = (
-            Electrical().get_trait(contructable_from_interface_list).from_interfaces(it)
-        )
-        self.IFs.inputs = list(
-            consume_iterator(
-                Electrical()
-                .get_trait(contructable_from_interface_list)
-                .from_interfaces,
-                it,
-            )
-        )
 
 
 class CD4011(Component):
